chore: pull panel and api definitions out to separate files
