+++ momentum/index.d.ts
@@ -4,37 +4,8 @@
  */
 
 /// <reference path="../shared/index.d.ts" />
-
-/** The map cache API. Exclusive to Momentum Mod! */
-declare namespace MapCacheAPI {
-	/** Gets the metadata for the current map */
-	function GetCurrentMapData(): unknown;
-
-	/** Gets the map data for a given mapID */
-	function GetMapDataByID(mapID: uint32): unknown;
-
-	/** Gets the current map's name. */
-	function GetMapName(): string;
-
-	/** Gets all the maps from the Map Cache */
-	function GetMaps(): unknown;
-
-	/** Returns true if the given mapID is queued for download */
-	function MapQueuedForDownload(mapID: uint32): boolean;
-}
-
-interface MapData {
-	name: string;
-	thumbnail: {
-		urlLarge: string;
-	}
-	mainTrack: {
-		difficulty: number;
-		numZones: number;
-		isLinear: boolean;
-	}
-	credits: { type: string, user: { alias: string } }[];
-}
+/// <reference path="./panels.d.ts" />
+/// <reference path="./apis.d.ts" />
 
 declare type TimerEvent = TimerEventEnum[keyof TimerEventEnum];
 declare interface TimerEventEnum {
